+++ tests/Datasets/LabeledTest.php
@@ -334,6 +334,13 @@ class LabeledTest extends TestCase
         $this->assertCount(3, $right);
     }
 
+    public function test_random_subset()
+    {
+        $subset = $this->dataset->randomSubset(3);
+
+        $this->assertCount(3, array_unique($subset->samples(), SORT_REGULAR));
+    }
+
     public function test_random_subset_with_replacement()
     {
         $subset = $this->dataset->randomSubsetWithReplacement(3);