+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Added early stopping window to Gradient Descent-based Learners
     - Changed early stopping behavior of MLP-based estimators
     - Added predict single sample method to Learner interface
+    - Changed method signature of random subset without replacement
 
 - 0.0.14-beta
     - Added feature importances to Gradient Boost